spec/elimination: Specify both roles for players
One set of tests depends on u2 having guerrilla. However, in some
circumstances u2 would not actually get it, because the other two
players would take up all copies of guerrilla. That means when u2 flips
a card that should be guerrilla, instead of u3 losing influence, u2 has
to lose influence. OOPS.

u1, u2, u3 are assigned post-initial-rotate, which means that u2 was the
last character to get cards (it's u3, u1, u2). u3 is sure to get one
Guerrilla, u1 is sure to get a second. If either of them get the third
Guerrilla, the test breaks.

The probability of u3 getting a guerrilla is 2/14.
The probability of u3 missing but u1 hitting is 12/14 * 1/12.
So that means 3/14 times the test would have broken.