next_decision: Check dead players before auto-complete
The other way is wrong because consider an each_live_card choice mixed
with a static choice (crime boss, disappear). See the comment on the
attached spec for more details.

Previously, next_decision checked auto-complete before dead players
because on-death resolution "decisions" have a dead player making a
"choice". Clearly that was bad design, which is why this was changed to
AutoDecision.

Now that dead players are by definition those who do not make choices,
we know we can check dead players before auto-complete.