Enqueue on death decisions at the end of the turn
If it were in the middle of a turn, a Lawyer may be able to use a dead
player's funds to pay off a Crime Boss, or a tax on a role the Lawyer
wants to claim. Instead, these all get processed at the end of a turn
during the :on_death phase specially created for the purpose.