spec: Add spec for blockable target all
This should have been there from the start since both target all actions
in G54 are in fact blockable.

It was not there before because blockability was tested in an orthogonal
manner, so it isn't really necessary to test a blockable target all.
(This spec does add a test for that, but that's not why it was added)

The real point of adding this spec is to test an interesting edge case.
Player A, with one influence, uses a blockable target all action.
Two players, B and C, are potentially affected.
Player B blocks.
Player A challenges the block.
Player B reveals the correct role.
Player A is now eliminated.
Player C *should* still be asked to block or not.
If player C chooses not to block, C should still be affected by the
action even after A has been eliminated.

So, this spec tests the case of being eliminated in the middle of your
action.